docking/requirements.txt
requests
websockets
Pillow
python-dotenv
